training/ui/damage: Avoid allocating new String in parse_anim_transform
`ResAnimationContent.name` is already assumed to be a valid
null-terminated C string so we can parse it in-place as a `CStr` rather
than calling `skyline::try_from_c_str` which allocates a new `String`
and copies then contents.

Additionally, make the following clause check the `Option` variant to
see if parsing was successful rather than perform a string comparison
against a sentinel value.